리뷰 사항 반영한 구조 변경
