+++ src/main/java/Main.java
@@ -1,16 +1,17 @@
-import domain.LottoSystem;
-import dto.LottosDto;
+import domain.LottoFactory;
+import domain.Lottos;
 import dto.ResultDto;
 import view.InputView;
 import view.OutputView;
 
 public class Main {
     public static void main(String[] args) {
-        LottoSystem lottoSystem = new LottoSystem();
-        LottosDto lottosDto = lottoSystem.makeLottos(InputView.purchaseAmount());
+        int price = InputView.purchaseAmount();
+        ResultDto resultDto = new ResultDto(price);
+        Lottos lottos = LottoFactory.makeLottos(price);
         OutputView outputView = new OutputView();
-        outputView.printLottoCount(lottosDto);
-        ResultDto resultDto = lottoSystem.checkLastWinningNumbers(InputView.lastAnswerNumbers());
+        outputView.printLottoCount(lottos);
+        resultDto.checkLastWinningNumbers(InputView.lastAnswerNumbers(), lottos);
         outputView.printResult(resultDto);
     }
 }
