+++ src/main/java/domain/Lottos.java
@@ -1,0 +1,19 @@
+package domain;
+
+import java.util.List;
+
+public class Lottos {
+    private List<Lotto> lottos;
+
+    public Lottos(List<Lotto> lottos) {
+        this.lottos = lottos;
+    }
+
+    public int size() {
+        return lottos.size();
+    }
+
+    public Lotto get(int index) {
+        return lottos.get(index);
+    }
+}
