+++ src/test/java/dto/ResultDtoTest.java
@@ -1,0 +1,56 @@
+package dto;
+
+import domain.Lotto;
+import domain.Lottos;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.Arrays;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class ResultDtoTest {
+    private ResultDto result;
+    private Lottos lottos;
+
+    @Before
+    public void setup() {
+        result = new ResultDto(1000);
+        lottos = new Lottos(Arrays.asList(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6))));
+    }
+
+    @Test
+    public void 로또를_산_갯수() {
+        assertThat(result.countOfLotto()).isEqualTo(1);
+    }
+
+    @Test
+    public void 못맞추거나_두개_이하로_맞췃을_때() {
+        result.checkLastWinningNumbers("7, 8, 9, 10, 11, 12", lottos);
+        assertThat(result.calRevenue()).isEqualTo(-100);
+    }
+
+    @Test
+    public void 세개_맞췃을_때() {
+        result.checkLastWinningNumbers("1, 2, 3, 10, 11, 12", lottos);
+        assertThat(result.calRevenue()).isEqualTo(400);
+    }
+
+    @Test
+    public void 네개_맞췃을_때() {
+        result.checkLastWinningNumbers("1, 2, 3, 4, 11, 12", lottos);
+        assertThat(result.calRevenue()).isEqualTo(4900);
+    }
+
+    @Test
+    public void 다섯개_맞췃을_때() {
+        result.checkLastWinningNumbers("1, 2, 3, 4, 5, 12", lottos);
+        assertThat(result.calRevenue()).isEqualTo(149900);
+    }
+
+    @Test
+    public void 여섯개_맞췃을_때() {
+        result.checkLastWinningNumbers("1, 2, 3, 4, 5, 6", lottos);
+        assertThat(result.calRevenue()).isEqualTo(199999900);
+    }
+}
